Make archive provenance checks conditional

+++ admin/validate-release.sh
@@ -10,6 +10,7 @@ fi
 if [ -z "$STAGE" ] ; then
   STAGE=dev # Alternatively release
 fi
+CheckProvenance=$(( $STAGE == "dev" ? 1 : 0 ))
 
 if [ -z "$BASE_DL" ] ; then
   BASE_DL=https://dist.apache.org/repos/dist/$STAGE/logging/log4cxx
@@ -24,24 +25,30 @@ fi
 test -d "$TEST_DIRECTORY" || mkdir "$TEST_DIRECTORY"
 cd "$TEST_DIRECTORY"
 
-WORKFLOW="package_code"
-echo "Downloading $WORKFLOW artifacts ..."
-# Get the latest Run ID
-RUN_ID=$(gh run list --repo apache/logging-log4cxx --workflow="$WORKFLOW.yml" --limit 1 --json databaseId --jq '.[0].databaseId')
-if [ $? -ne 0 ]; then
-  echo "Failed to find a GitHub $WORKFLOW run id"
-  exit 1
-fi
-# Download the artifacts
-test -d release_files && rm -rf release_files
-gh run download --repo apache/logging-log4cxx "$RUN_ID"
-if [ $? -ne 0 ] || [ ! -d release_files ]; then
-  echo "Failed to download GitHub $WORKFLOW run $RUN_ID artifacts"
-  exit 1
-fi
-if [ ! -f "release_files/$ARCHIVE.tar.gz.sha512" ] ; then
-  echo "$ARCHIVE.tar.gz.sha512 not found in GitHub $WORKFLOW run $RUN_ID artifacts"
-  exit 1
+if $CheckProvenance ; then
+  if gh --version >> /dev/null ; then
+    WORKFLOW="package_code"
+    echo "Downloading $WORKFLOW artifacts ..."
+    # Get the latest Run ID
+    RUN_ID=$(gh run list --repo apache/logging-log4cxx --workflow="$WORKFLOW.yml" --limit 1 --json databaseId --jq '.[0].databaseId')
+    if [ $? -ne 0 ]; then
+      echo "Failed to find a GitHub $WORKFLOW run id"
+      exit 1
+    fi
+    # Download the artifacts
+    test -d release_files && rm -rf release_files
+    gh run download --repo apache/logging-log4cxx "$RUN_ID"
+    if [ $? -ne 0 ] || [ ! -d release_files ]; then
+      echo "Failed to download GitHub $WORKFLOW run $RUN_ID artifacts"
+      exit 1
+    fi
+    if [ ! -f "release_files/$ARCHIVE.tar.gz.sha512" ] ; then
+      echo "$ARCHIVE.tar.gz.sha512 not found in GitHub $WORKFLOW run $RUN_ID artifacts"
+      exit 1
+    fi
+  else
+    echo "GitHub CLI program (gh) is not available - provenance checks will be skipped"
+  fi
 fi
 
 FULL_DL="$BASE_DL/$VERSION/$ARCHIVE"
@@ -59,12 +66,14 @@ for ARCHIVE_TYPE in "tar.gz" "zip" ; do
   echo "Validating signature..."
   gpg --verify "$ARCHIVE.$ARCHIVE_TYPE.asc" || exit $?
 
-  echo "Checking provenance ..."
-  if diff {,release_files/}$ARCHIVE.$ARCHIVE_TYPE.sha512 ; then
-    echo "$ARCHIVE.$ARCHIVE_TYPE is from a GitHub workflow"
-  else
-    echo "$ARCHIVE.$ARCHIVE_TYPE is not from a GitHub workflow"
-    exit 1
+  if [ -f release_files/$ARCHIVE.$ARCHIVE_TYPE.sha512 ] ; then
+    echo "Checking provenance ..."
+    if diff {,release_files/}$ARCHIVE.$ARCHIVE_TYPE.sha512 ; then
+      echo "$ARCHIVE.$ARCHIVE_TYPE is from a GitHub workflow"
+    else
+      echo "$ARCHIVE.$ARCHIVE_TYPE is not from a GitHub workflow"
+      exit 1
+    fi
   fi
 done
